feat(pool): add dynamic concurrency
Default to 1–8 sessions and async workers with a 60-second idle
timeout. Preserve fixed-count constructors and validate policies
before opening storage.

BREAKING CHANGE: replace CommandStrands::count() with max().

Closes #34

+++ src/nonblocking/shutdown.rs
@@ -25,10 +25,10 @@ pub(super) struct Shutdown {
 }
 
 impl Shutdown {
-    pub(super) fn new(workers: u32, reply: oneshot::Sender<Result<()>>) -> Self {
+    pub(super) fn new(workers: usize, reply: oneshot::Sender<Result<()>>) -> Self {
         Self {
-            remaining: AtomicUsize::new(workers as usize),
-            handles: Mutex::new(Vec::with_capacity(workers as usize)),
+            remaining: AtomicUsize::new(workers),
+            handles: Mutex::new(Vec::with_capacity(workers)),
             reply: Mutex::new(Some(reply)),
         }
     }
